IECoreAppleseed::Renderer: added support for interactive non-editable final rendering.
